xiulian/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xiulian"
version = "1.0.0"
description = "轻量级符号推理引擎 - 非Transformer架构的AI系统"
license = {text = "GPL-3.0"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "networkx>=3.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "flask>=2.3.0",
    "PySide6>=6.5.0",
    "pyahocorasick>=2.0.0",
]

[project.scripts]
xiulian = "xiulian.cli:main"

[tool.setuptools.packages.find]
include = ["xiulian*"]